config-editor로 파일 분리
